[refactor](storage) move encode_field_as_key helpers into key_coder.h
The Field-to-key encoding helpers and the dispatch X-macro fit
naturally next to KeyCoder rather than in a stand-alone header,
since they are thin wrappers around KeyCoder calls. Inline them
into storage/key_coder.h and remove storage/field_key_encoder.h.
