+++ be/src/storage/index/inverted/inverted_index_reader.cpp
@@ -45,7 +45,6 @@
 #include "runtime/runtime_profile.h"
 #include "runtime/runtime_state.h"
 #include "storage/field.h"
-#include "storage/field_key_encoder.h"
 #include "storage/index/index_file_reader.h"
 #include "storage/index/index_reader_helper.h"
 #include "storage/index/inverted/analyzer/analyzer.h"
